requirements.txt
boto3
flask
psycopg2-binary